FUZZY C MEANS AND CATBOOST ALGORITHM FOR FAST AND OPTIMIZED VIDEO CODING
In today's modern technologies video encoding is important because it allows us to more easily transmit video content over the internet. In video streaming, encoding is crucial because the compressing of the raw video reduces the bandwidth making it easier to transmit, while still maintaining a good quality of experience for end viewers. In this paper, we propose a machine learning based Fuzzy C Means and CatBoost algorithm to reduce the coding loss and encoding time on HEVC for an optimized video.We compare the input video and the optimized video using two parameters PSNR(peak signal-to-noise ratio) and SSIM(structural similarity index measure) to measure the quality of imperceptibility